requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-menu==0.1.13
django-sitetree==1.14.0
django-taggit==1.3.0
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
Pillow==7.1.1
psycopg2-binary==2.8.6
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.10
whitenoise==5.2.0
